Auslagerung von Methoden in andere Dateien zur Verbesserung der Übersichtlichkeit
